change the page to render and modificated the sanitary register
